refactor: remove dl from MetricItem
