+++ src/components/budgets/BudgetCard.tsx
@@ -16,6 +16,7 @@ import IconButton from "@/components/ui/IconButton"
 import Label from "@/components/ui/Label"
 import Link from "@/components/ui/Link"
 import { Menu, MenuTrigger, MenuItem } from "@/components/ui/Menu"
+import MetricItem from "@/components/ui/MetricItem"
 import { Category, Color } from "@/data-access/lookups"
 import { currencyFormatter } from "@/lib/utils"
 
@@ -97,29 +98,18 @@ export default function BudgetCard({
         )}
 
         <div className="grid grid-cols-2">
-          <div className="grid grid-cols-[auto_1fr] gap-x-4 gap-y-1">
-            <span
-              className="row-span-2 h-full w-1 rounded-lg"
-              style={{ backgroundColor: budget.color.value }}
-            />
-            <p className="text-grey-500 text-xs leading-normal font-normal">
-              Spent
-            </p>
-            <p className="text-grey-900 text-sm leading-normal font-bold">
-              {currencyFormatter.format(Number(budget.totalSpent))}
-            </p>
-          </div>
-          <div className="grid grid-cols-[auto_1fr] gap-x-4 gap-y-1">
-            <span className="bg-beige-100 row-span-2 h-full w-1 rounded-lg" />
-            <p className="text-grey-500 text-xs leading-normal font-normal">
-              Free
-            </p>
-            <p className="text-grey-900 text-sm leading-normal font-bold">
-              {currencyFormatter.format(
-                remainingAmount < 0 ? 0 : remainingAmount
-              )}
-            </p>
-          </div>
+          <MetricItem
+            label="Spent"
+            color={budget.color.value}
+            value={currencyFormatter.format(Number(budget.totalSpent))}
+          />
+          <MetricItem
+            label="Free"
+            color="#f8f4f0"
+            value={currencyFormatter.format(
+              remainingAmount < 0 ? 0 : remainingAmount
+            )}
+          />
         </div>
 
         <div className="bg-beige-100 grid gap-5 rounded-xl p-4 md:p-5">
